requirements/dev.txt
-r base.txt
mysqlclient==1.3.6
django-extensions
django-debug-toolbar
